Provide rough draft documentation for the API
This documentation is very rough. It should allow us to get started, but
I'm sure that people will find faults and I will need to redo portions
of it. The idea is to get something out and polish it as we add to it
and correct functionality in the calls.

TECHOPS-921