Fix the change password token validation endpoint+ (#104)
While fixing that, I realized that the password calls a basically
internal calls and should be documented on the internal document. If a
customer needs to go through the password process, they should use the
UI.

And because of that, I reorganized the endpoints into the organization
section.

So basically a bunch of cleanup that stemmed from one tiny incorrect
call.